Adding a StandardsRegExt record for VOEvent.
This ought to be mailed to the RofR as soon as we're happy with it, and
then updated when a new version is issued.

I've taken the liberty to include a few standard keys required in
the proposal for registering VOEvent streams that'll come in a separate
PR.  They don't hurt if if that's rejected (but probably should still
be taken out if that happens).